feat: add all tables with relations

+++ src/Domain/Entities/Service.cs
@@ -8,8 +8,9 @@ public class Service
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
-    public decimal Price { get; set; }
+    public float Price { get; set; }
     public int Duration { get; set; } // Duration in hours
+    
     public virtual List<Skill> Skills { get; set; } = [];
     public virtual List<Appointment>? Appointments { get; set; } = [];
 }
